Move init_db script from tests, add Contributing info
