--- demos/README.rst
@@ -1,2 +1,0 @@
-aiohttp demos
-=============
