+++ demos/polls/init_db.py
